requirements.txt
docker==5.0.3
sanic==21.12.1
aiofiles==0.8.0
redis==4.1.4
pyjwt==2.3.0